requirements.txt
asgiref==3.8.1
contourpy==1.3.1
crispy==0.8.0
crispy-bootstrap5==2024.10
cycler==0.12.1
Django==5.1.6
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fabio==2024.9.0
fonttools==4.56.0
greenlet==3.1.1
h5py==3.13.0
hdf5plugin==5.0.0
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psycopg==3.2.5
psycopg2==2.9.10
PyJWT==2.9.0
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
silx==2.2.1
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
xraydb==4.5.6
